requirements.txt
bdfparser==2.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
colorzero==2.0
dateutils==0.6.12
frozendict==2.4.4
geojson==2.5.0
gevent==24.2.1
gpiozero==2.0.1
greenlet==3.0.3
html5lib==1.1
idna==3.7
libsixel-python==0.5.0
lxml==5.2.2
multitasking==0.0.11
numpy==2.0.0
pandas==2.2.2
peewee==3.17.6
pillow==10.4.0
platformdirs==4.2.2
pygame==2.6.0
pyowm==3.3.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.7
requests==2.32.3
RGBMatrixEmulator==0.11.6
scipy==1.14.0
setuptools==70.1.1
six==1.16.0
soupsieve==2.5
spotipy==2.24.0
tornado==6.4.1
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
websocket==0.2.1
wheel==0.43.0
yfinance==0.2.40
zope.event==5.0
zope.interface==6.4.post2
lgpio==0.2.2.0
pyyaml==6.0.2